Prediction ValueRange: fix possible NPE if value is null
- In my refactoring of the Prediction class I missed handling of a situation where a NullPointerException could happen.
- This leads to the problem that - if no local historic values are existing, e.g. after system outage or at IBN, the prediction fails with a NullPointerException.
